[#102] Move the autoloader back to Familia::Features
