+++ docs/migrating/v2.0.0-pre11.md
@@ -124,7 +124,7 @@ end
 
 class Customer < Familia::Horreum
   module Features
-    include Familia::Autoloader
+    include Familia::Features::Autoloader
     # Automatically discovers and loads all *.rb files from customer/features/
   end
 end
@@ -210,7 +210,7 @@ If you have project-specific features, set up auto-loading:
 module YourProject
   class ModelName < Familia::Horreum
     module Features
-      include Familia::Autoloader
+      include Familia::Features::Autoloader
     end
   end
 end
