+++ docs/migrating/v2.0.0-pre13.md
@@ -68,7 +68,7 @@ Common issues:
 
 The Feature Autoloading System consists of two key components:
 
-### Familia::Autoloader
+### Familia::Features::Autoloader
 A utility module providing shared file loading functionality:
 - Handles Dir.glob pattern matching and file loading
 - Provides consistent debug logging across all autoloading scenarios
